requirements.txt
streamlit==1.28.1
mysql-connector-python==8.0.33
pandas==2.0.3
plotly==5.17.0
numpy==1.24.3
scipy==1.11.0